+++ src/app/record/pages/my-orcid/my-orcid.component.ts
@@ -74,7 +74,6 @@ export class MyOrcidComponent implements OnInit, OnDestroy {
   regionActivities = $localize`:@@shared.activities:Activities`
   readyForIndexing: boolean
   fragment: string
-  newRecordHeaderTogglz = false
   newRecordHeaderWithSummaryTogglz: boolean
 
   constructor(
@@ -224,10 +223,6 @@ export class MyOrcidComponent implements OnInit, OnDestroy {
                 this.userRecord
               )
             }
-          }
-
-          if (togglz.messages['NEW_RECORD_HEADER'] === 'true') {
-            this.newRecordHeaderTogglz = true
           }
 
           if (togglz.messages['CRAZY_EGG'] === 'true') {